phpPgAdmin for DSM 6.1 || Read the description!!!
The install is not perfect. Because I'm no synology/linux guru. 

But it is running well if you do as follows:
Login with putty or bash and create pg admin:
sudo -i
login postgres:
su postgres
psql
list user and roles and add user 'admin' with your password!
\du
create admin user in postgres:
CREATE ROLE admin WITH PASSWORD 'xyz' SUPERUSER CREATEDB CREATEROLE LOGIN;
\du
\q
exit

RUN install
Restart your Synology.

If it is not running, reboot synology or go to church.:-D
There is a very nice workflow script how to do this manually: https://forum.synology.com/enu/viewtopic.php?t=123824

phpPgAdmin is running as "http" user. That should be secure. But as mentioned, I don't know. 

And don't forget, the answer is 42,
and thanks for all the fish,
enzolino